requirements.txt
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
atari-py==0.1.1
azure==1.0.3
azure-common==1.1.8
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==2.0.0
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==2.0.0
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
Babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
-e git+https://github.com/openai/baselines.git@6a3cbb4bc57633b2e0d6aa1cb9556c2b4a0b6989#egg=baselines
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.1
bokeh==0.12.10
boto==2.48.0
boto3==1.5.2
botocore==1.8.16
Bottleneck==1.2.1
bz2file==0.98
certifi==2017.11.5
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
cmake==0.8.0
colorama==0.3.9
conda==4.3.30
conda-build==3.0.27
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.0.3
cycler==0.10.0
cymem==1.31.2
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.3
datashape==0.5.4
de-core-news-sm==2.0.0
decorator==4.1.2
dill==0.2.7.1
distributed==1.19.1
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.12
Flask==0.12.2
Flask-Cors==3.0.3
ftfy==4.4.3
future==0.16.0
gensim==3.2.0
gevent==1.2.2
glob2==0.5
gmpy2==2.0.8
greenlet==0.4.12
-e git+https://github.com/openai/gym@b5576dc23a5fcad0733042ab2ad440200ebb6209#egg=gym
h5py==2.7.0
heapdict==1.0.0
html5lib==1.0b10
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipdb==0.10.3
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.3
joblib==0.11
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
Keras==2.1.0
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lxml==4.1.0
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mistune==0.7.4
mpi4py==3.0.0
mpmath==0.19
msgpack-numpy==0.4.1
msgpack-python==0.4.8
mujoco-py==0.5.7
multipledispatch==0.4.9
murmurhash==0.28.0
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0+10.g143f70e90
numexpr==2.6.2
numpy==1.13.3
numpydoc==0.7.0
odo==0.5.1
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib==1.0.1
pathlib2==2.3.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
pkginfo==1.4.1
plac==0.9.6
ply==3.10
preshed==1.0.0
progressbar2==3.34.3
prompt-toolkit==1.0.15
protobuf==3.4.0
psutil==5.4.0
ptyprocess==0.5.2
py==1.4.34
pycodestyle==2.3.1
pycosat==0.6.2
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.6.0
pyglet==1.3.0
Pygments==2.2.0
pylint==1.7.4
pyodbc==4.0.17
PyOpenGL==3.1.0
pyOpenSSL==17.2.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.6.1
python-utils==2.2.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
regex==2017.4.5
requests==2.18.4
rope==0.10.5
ruamel-yaml==0.11.14
s3transfer==0.1.12
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.5
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
spacy==2.0.3
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
SQLAlchemy==1.1.13
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
thinc==6.10.1
toolz==0.8.2
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.2
typing==3.6.2
ujson==1.35
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zict==0.1.3
zmq==0.0.0
